API_Face/requirements.txt
Flask==3.0.3
kafka==1.3.5
numpy==1.24.4
onnxruntime_gpu==1.16.0
opencv_contrib_python==4.10.0.84
# opencv_python==4.9.0.80
opencv_python_headless==4.10.0.84
pandas==2.0.3
# psycopg2==2.9.10
pymongo==4.10.1
# PyYAML==5.4.1
# PyYAML==6.0.2
Requests==2.32.3
scikit-image
networkx==2.8.8
torch==2.4.0+cu118
-f https://download.pytorch.org/whl/torch_stable.html

